feat(fastly): use edge dictionaries for dynamic redirects
we continue to create one condition and header for each regular line in the redirects config, but dynamic redirects that are based on spreadsheet data will be handled differently. Each spreadsheet is turned into two edge dictionaries, one for permanent and one for temporary redirects. Then two conditions and headers are created that perform the dictionary lookup. This brings the overall number of conditions down to managable levels, even for large redirect tables.